+++ src/commands/index.ts
@@ -2,5 +2,7 @@ import MainCommand from "./main";
 import LoginCommand from "./login";
 import ResetCommand from "./reset";
 import ConfigCommand from "./config";
+import CompletionsCommand from "./completions";
 
-export { MainCommand, LoginCommand, ResetCommand, ConfigCommand };
+export { MainCommand, LoginCommand, ResetCommand, CompletionsCommand, ConfigCommand };
+